requirements.txt
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
pyknow==1.7.0
python-firebase==1.2
gunicorn
